Move callback handlers back into FormWrapper, outside of context scope
